Add placeholder of flow answering after select business. Add handling flow for restarting the bot.
Handling flow needs to be added thus checking or /start to set the appropriate db record to reset to Normal state.
This is due to deleting the chat, and go back again. The keyboard state, and user's state as saved in db should be reset. Thus handling resolves this.

Add /startover command to start over during the conversation.
This command is not used just yet, but in case we need it.

Also updated gist command link.